backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.15.0
mediapipe==0.10.7
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
openai==1.3.5
requests==2.31.0
replicate==0.25.1
httpx==0.25.2
